(fix) Refactor getRepoOwners into a helper function
