+++ server/controllers/commitController.js
@@ -3,6 +3,7 @@ const db = require('../db/database.js').db;
 const pgp = require('../db/database.js').pgp;
 const token = require('../config/github.config').token;
 const rp = require('request-promise');
+const getRepoOwners = require('../helpers/getRepoOwners');
 
 // GET at '/api/v1/users/:id/commits'
 exports.retrieveCommits = function(req, res) {
@@ -77,7 +78,6 @@ exports.updateCommits = function(req, res) {
           .then(commits => {
             // send back the updated commits
             console.log('Successfully patched commits for userId: ' + userId);
-            console.log('sending commits: ', commits);
             res.send(commits);
           })
           .catch(error => {
@@ -96,80 +96,6 @@ exports.updateCommits = function(req, res) {
       res.status(500).send;
     }); 
   };            
-  
-  // get the owner information for each repo
-  // this will allow us to set up the GitHub GET request
-  var getRepoOwners = callback => {
-    // get the user for this queryId
-    db.one('SELECT * from $1~ ' +
-      'WHERE $2~=$3',
-      ['users', 'id', queryId])
-      // get the repos for this user
-      .then(user => {
-        db.any('SELECT r.id, r.owner_id, r.name ' +
-          'FROM users_repos ur ' +
-          'INNER JOIN repos r ' +
-          'ON r.id = ur.repo_id ' +
-          'WHERE ur.user_id=$1',
-          [queryId])
-          // set up a GET request for each of this user's repos
-          .then(repos => {
-            var repoCountGetRepoOwners = 0;
-            var totalRepos = repos.length;
-            var ownerName = user.username;
-            var repoName = '';
-            var repoOwners = [];
-            // set up a hash of each repo's ownership info
-            repos.forEach(repo => {
-              db.oneOrNone('SELECT * FROM $1~ ' +
-                'WHERE $2~=$3',
-                ['orgs', 'id', repo.owner_id])
-                .then(org => {
-                  if (org !== null) {
-                    repoOwners.push({
-                      repoId: repo.id,
-                      repoName: repo.name,
-                      userName: user.username,
-                      userId: user.id,
-                      orgId: org.id,
-                      orgName: org.orgname,
-                      repoOwnerId: repo.owner_id
-                    });
-                  } else {
-                    repoOwners.push({
-                      repoId: repo.id,
-                      repoName: repo.name,
-                      userName: user.username,
-                      userId: user.id,
-                      orgId: null,
-                      orgName: null,
-                      repoOwnerId: repo.owner_id
-                    });
-                  }
-                })
-                .then(data => {
-                  repoCountGetRepoOwners++;
-                  if (repoCountGetRepoOwners === totalRepos) {
-                    callback(repoOwners);
-                  }
-                })  
-                .catch(error => {
-                  console.error('Error selecting orgs: ', error);
-                  res.status(500).send;
-                });
-            });
-          })  
-          .catch(error => {
-            console.error('Error selecting repos: ', error);
-            res.status(500).send;
-          });       
-      })
-      .catch(error => {
-        console.error('Error selecting user: ', error);
-        res.status(500).send;
-      });
-  };
-  
   
   var getCommitsFromGitHub = (repoOwners) => {
     var totalRepos = repoOwners.length;
@@ -240,6 +166,6 @@ exports.updateCommits = function(req, res) {
   
   // CALL HELPER FUNCTIONS
   
-  getRepoOwners(getCommitsFromGitHub);
+  getRepoOwners(queryId, getCommitsFromGitHub);
   
 };
